+++ src/main/java/org/nusco/narjillos/core/utilities/Configuration.java
@@ -120,9 +120,9 @@ public class Configuration {
 	public static final double ECOSYSTEM_EGGS_DENSITY_PER_BLOCK = getDouble("ecosystem", "eggs_density_per_block");
 	public static final double ECOSYSTEM_FOOD_DENSITY_PER_BLOCK = getDouble("ecosystem", "food_density_per_block");
 	public static final double ECOSYSTEM_MAX_FOOD_DENSITY_PER_1000_BLOCK = getDouble("ecosystem", "max_food_density_per_1000_blocks");
-	public static final double ECOSYSTEM_FOOD_RESPAWN_AVERAGE_INTERVAL_PER_BLOCK = getInt("ecosystem", "food_respawn_average_interval_per_block");
+	public static final int ECOSYSTEM_FOOD_RESPAWN_AVERAGE_INTERVAL_PER_BLOCK = getInt("ecosystem", "food_respawn_average_interval_per_block");
 	public static final int ECOSYSTEM_UPDATE_FOOD_TARGETS_INTERVAL = getInt("ecosystem", "update_food_targets_interval");
-	public static final int ECOSYSTEM_INITIAL_ELEMENT_LEVEL = getInt("ecosystem", "initial_element_level");
+	public static final double ECOSYSTEM_INITIAL_ELEMENT_LEVEL = getDouble("ecosystem", "initial_element_level");
 	public static final int ECOSYSTEM_CATALYST_LEVEL = getInt("ecosystem", "catalyst_level");
 
 	// experiment
